+++ src/SparkFun_VEML6030_Ambient_Light_Sensor.h
@@ -172,32 +172,32 @@ class SparkFun_Ambient_Light
     // REG0x02, bits[15:0]
     // This function sets the lower limit for the Ambient Light Sensor's interrupt. 
     // It takes a lux value as its paramater.
-    void setIntLowThresh(uint32_t luxVal);
+    void setIntLowThresh(float luxVal);
     
     // REG0x02, bits[15:0]
     // This function reads the lower limit for the Ambient Light Sensor's interrupt. 
-    uint32_t readLowThresh();
+    float readLowThresh();
 
     // REG0x01, bits[15:0]
     // This function sets the upper limit for the Ambient Light Sensor's interrupt. 
     // It takes a lux value as its paramater.
-    void setIntHighThresh(uint32_t luxVal);
+    void setIntHighThresh(float luxVal);
 
     // REG0x01, bits[15:0]
     // This function reads the upper limit for the Ambient Light Sensor's interrupt. 
-    uint32_t readHighThresh();
+    float readHighThresh();
 
     // REG[0x04], bits[15:0]
     // This function gets the sensor's ambient light's lux value. The lux value is
     // determined based on current gain and integration time settings. If the lux
     // value exceeds 1000 then a compensation formula is applied to it. 
-    uint32_t readLight();
+    float readLight();
 
     // REG[0x05], bits[15:0]
     // This function gets the sensor's ambient light's lux value. The lux value is
     // determined based on current gain and integration time settings. If the lux
     // value exceeds 1000 then a compensation formula is applied to it. 
-    uint32_t readWhiteLight();
+    float readWhiteLight();
 
   private:
 
@@ -207,21 +207,21 @@ class SparkFun_Ambient_Light
     // "Illumination values higher than 1000 lx show non-linearity. This
     // non-linearity is the same for all sensors, so a compensation forumla..."
     // etc. etc. 
-    uint32_t _luxCompensation(uint32_t _luxVal);
+    float _luxCompensation(float _luxVal);
 
     // The lux value of the Ambient Light sensor depends on both the gain and the
     // integration time settings. This function determines which conversion value
     // to use by using the bit representation of the gain as an index to look up
     // the conversion value in the correct integration time array. It then converts 
     // the value and returns it.  
-    uint32_t _calculateLux(uint16_t _lightBits);
+    float _calculateLux(uint16_t _lightBits);
 
     // This function does the opposite calculation then the function above. The interrupt
     // threshold values given by the user are dependent on the gain and
     // intergration time settings. As a result the lux value needs to be
     // calculated with the current settings and this function accomplishes
     // that.  
-    uint16_t _calculateBits(uint32_t _luxVal);
+    uint16_t _calculateBits(float _luxVal);
 
     // This function writes to a 16 bit register. Paramaters include the register's address, a mask 
     // for bits that are ignored, the bits to write, and the bits' starting
